[jit] Add namespacing for ScriptClasses
All classes now get a namespace. The compiler frontend now knows what namespace
it's in (if any) and can only look up or define classes in that namespace when
emitting IR.

In C++, modules are now imported into a private namespace, which
prevents multiple models that define the same classname from clobbering
each other.

In Python, everything goes in a "python" namespace for now. This can
easily be changed to take __module__ or something if needed.

Not implemented: any form of shared libraries or nested namespaces.
Shouldn't be too hard to implement on top of what we have here, but I'm
keeping it simple for now.

gh-metadata: pytorch pytorch 18296 gh/suo/14/head